XMPP:ServiceResolver: Fix a potential bug if an invalid port was specified
Previously this invalid port could have been returned via resultReady(), thus
 early santity checking seems better.

+++ src/irisnet/corelib/netnames.h
@@ -590,7 +590,7 @@ public:
 	 * \param host Hostname to lookup
 	 * \param port Port to signal via resultReady (for convenience)
 	 */
-	void start(const QString &host, int port);
+	void start(const QString &host, quint16 port);
 	/*!
 	 * Start an indirect (SRV) lookup for the service
 	 * \param service Service type, like "ssh" or "ftp"
